+++ commons-persistence/src/main/java/org/dependencytrack/persistence/model/CsafDocumentEntity.java
@@ -20,8 +20,6 @@ package org.dependencytrack.persistence.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
@@ -38,9 +36,8 @@ import java.io.Serializable;
 public class CsafDocumentEntity implements Serializable {
 
     @Id
-    @Column(name = "ENTRYID")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long entryId;
+    @Column(name = "ID")
+    private String id;
 
     @Column(name = "NAME")
     private String name;
@@ -75,12 +72,12 @@ public class CsafDocumentEntity implements Serializable {
         this.url = url;
     }
 
-    public long getEntryId() {
-        return entryId;
+    public String getId() {
+        return id;
     }
 
-    public void setEntryId(long entryId) {
-        this.entryId = entryId;
+    public void setId(String id) {
+        this.id = id;
     }
 
     public String getName() {
