+++ commons-persistence/src/main/java/org/dependencytrack/persistence/model/CsafSourceEntity.java
@@ -22,6 +22,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 
 import java.io.Serializable;
 import java.time.Instant;
@@ -37,8 +39,9 @@ import java.time.Instant;
 public class CsafSourceEntity implements Serializable {
 
     @Id
-    @Column(name = "ID")
-    private String id;
+    @Column(name = "ENTRYID")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long entryId;
 
     @Column(name = "NAME")
     private String name;
@@ -76,12 +79,12 @@ public class CsafSourceEntity implements Serializable {
         this.url = url;
     }
 
-    public String getId() {
-        return id;
+    public long getEntryId() {
+        return entryId;
     }
 
-    public void setId(String id) {
-        this.id = id;
+    public void setEntryId(long entryId) {
+        this.entryId = entryId;
     }
 
     public String getName() {
